backend/requirements.txt
fastapi
uvicorn
faster_whisper
librosa
sounddevice
soundfile
pydub
numpy
coqui-tts
scipy
pandas
pyannote.audio
pyyaml